Change required command-line argument structure.
This is an important change.  Previously, the code used to require 4
command-line arguments.  Now it requires only one (`FESC').  Variables
that used to take the other three command-line arguments have now been
assigned default values.  (These variables cannot simply be removed
from the code because they are somewhat deeply embedded, e.g., in
pigmd.f90.)  This change was overdue.  I wasted last two weeks on this
because this was not done before!